requirements.txt
fire
loguru
python-nmap
